Account for not awarding skill level when sending email (fixes #72)
When sending the email to the member to say that their skill level has
been processed, the system looked up the name of the skill level from
the hardcoded array. This didn't account for when the skill level wasn't
awarded, throwing an error. As the text is only shown when a level is
awarded, a check has been added that only looks up the name of the skill
level when it's been awarded.