Explicitly set uncategorised training skill category as null (fixed #70)
When updating training categories, the `category_id` field is updated -
when this set to 'Uncategorised', although Laravel stores it as `null`,
using the clean() and the default get() method returns an empty string.
This is not the same as NULL as far as MySQL is concerned, and so the
foreign key constraint was failing. This has been rectified by explicitly
setting the key to `null` when the `category_id` value is empty.